requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
black==19.10b0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0
django-anymail==7.0.0
django-cors-headers==3.2.0
django-extensions==2.2.5
django-filter==2.2.0
django-model-utils==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
pathspec==0.6.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
redis==3.3.11
regex==2019.12.9
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
zipp==0.6.0
